requirements.txt
OpenDAP
mutagen==1.41.1
python-vlc==3.0.7110